Move TestStateUtils, remove platform-core testFixtures
