+++ hedera-node/hedera-app/build.gradle.kts
@@ -31,7 +31,6 @@ testModuleInfo {
     requires("com.hedera.node.config.test.fixtures")
     requires("com.swirlds.merkledb")
     requires("com.swirlds.config.extensions.test.fixtures")
-    requires("com.swirlds.platform.core.test.fixtures")
     requires("com.swirlds.state.api.test.fixtures")
     requires("com.swirlds.state.impl.test.fixtures")
     requires("com.swirlds.base.test.fixtures")
