+++ platform-sdk/consensus-gossip-impl/build.gradle.kts
@@ -10,7 +10,6 @@ description = "Default Consensus Gossip Implementation"
 testModuleInfo {
     requires("com.swirlds.base.test.fixtures")
     requires("com.swirlds.platform.core")
-    requires("com.swirlds.platform.core.test.fixtures")
     requires("com.swirlds.config.extensions")
     requires("com.swirlds.config.extensions.test.fixtures")
     requires("org.hiero.base.crypto.test.fixtures")
