+++ platform-sdk/consensus-hashgraph-impl/src/main/java/module-info.java
@@ -4,19 +4,16 @@ import org.hiero.consensus.hashgraph.impl.DefaultHashgraphModule;
 
 module org.hiero.consensus.hashgraph.impl {
     exports org.hiero.consensus.hashgraph.impl to
-            com.swirlds.platform.core.test.fixtures,
             org.hiero.consensus.gossip.impl.test.fixtures,
             org.hiero.consensus.gui,
             org.hiero.consensus.hashgraph.impl.test.fixtures,
             org.hiero.consensus.network.simulation,
             org.hiero.consensus.network.simulation.test.fixtures;
     exports org.hiero.consensus.hashgraph.impl.consensus to
-            com.swirlds.platform.core.test.fixtures,
             org.hiero.consensus.gui,
             org.hiero.consensus.hashgraph.impl.test.fixtures,
             org.hiero.consensus.pcli;
     exports org.hiero.consensus.hashgraph.impl.linking to
-            com.swirlds.platform.core.test.fixtures,
             org.hiero.consensus.gui,
             org.hiero.consensus.hashgraph.impl.test.fixtures;
     exports org.hiero.consensus.hashgraph.impl.metrics to
