+++ platform-sdk/consensus-reconnect-impl/build.gradle.kts
@@ -10,7 +10,6 @@ mainModuleInfo { annotationProcessor("com.swirlds.config.processor") }
 
 testModuleInfo {
     requires("com.swirlds.base.test.fixtures")
-    requires("com.swirlds.platform.core.test.fixtures")
     requires("com.swirlds.config.extensions.test.fixtures")
     requires("com.swirlds.state.impl.test.fixtures")
     requires("com.swirlds.merkledb.test.fixtures")
