+++ platform-sdk/consensus-state-management/build.gradle.kts
@@ -12,7 +12,6 @@ testModuleInfo {
     requires("com.swirlds.config.extensions.test.fixtures")
     requires("com.swirlds.merkledb.test.fixtures")
     requires("com.swirlds.platform.core")
-    requires("com.swirlds.platform.core.test.fixtures")
     requires("com.swirlds.state.impl.test.fixtures")
     requires("org.hiero.base.crypto.test.fixtures")
     requires("org.hiero.base.utility.test.fixtures")
