+++ platform-sdk/swirlds-cli/build.gradle.kts
@@ -10,7 +10,6 @@ mainModuleInfo { runtimeOnly("org.hiero.consensus.pces.noop.impl.test.fixtures")
 testModuleInfo {
     requires("com.swirlds.common.test.fixtures")
     requires("com.swirlds.config.extensions.test.fixtures")
-    requires("com.swirlds.platform.core.test.fixtures")
     requires("org.hiero.consensus.event.stream.test.fixtures")
     requires("org.hiero.base.utility.test.fixtures")
     requires("org.hiero.consensus.pcli")
