+++ platform-sdk/swirlds-platform-core/build.gradle.kts
@@ -3,7 +3,6 @@ plugins {
     id("org.hiero.gradle.module.library")
     id("org.hiero.gradle.feature.publish-artifactregistry")
     id("org.hiero.gradle.feature.benchmark")
-    id("org.hiero.gradle.feature.test-fixtures")
     id("org.hiero.gradle.feature.test-timing-sensitive")
 }
 
@@ -26,7 +25,6 @@ mainModuleInfo {
 
 jmhModuleInfo {
     requires("com.swirlds.platform.core")
-    requires("com.swirlds.platform.core.test.fixtures")
     requires("com.hedera.node.hapi")
     requires("org.hiero.consensus.model.test.fixtures")
     requires("org.hiero.consensus.pces")
@@ -40,7 +38,6 @@ testModuleInfo {
     requires("com.swirlds.base.test.fixtures")
     requires("com.swirlds.common.test.fixtures")
     requires("com.swirlds.platform.core")
-    requires("com.swirlds.platform.core.test.fixtures")
     requires("com.swirlds.config.extensions.test.fixtures")
     requires("com.swirlds.state.api.test.fixtures")
     requires("com.swirlds.state.impl.test.fixtures")
@@ -65,7 +62,6 @@ testModuleInfo {
 timingSensitiveModuleInfo {
     requires("com.swirlds.metrics.api")
     requires("com.swirlds.platform.core")
-    requires("com.swirlds.platform.core.test.fixtures")
     requires("com.swirlds.state.impl")
     requires("org.hiero.base.utility.test.fixtures")
     requires("org.hiero.consensus.metrics")
