--- platform-sdk/swirlds-platform-core/src/testFixtures/java/com/swirlds/platform/test/fixtures/config/ConfigUtils.java
@@ -1,30 +1,0 @@
-// SPDX-License-Identifier: Apache-2.0
-package com.swirlds.platform.test.fixtures.config;
-
-import com.swirlds.config.api.Configuration;
-import com.swirlds.config.api.ConfigurationBuilder;
-import com.swirlds.config.extensions.sources.SimpleConfigSource;
-import com.swirlds.merkledb.config.MerkleDbConfig;
-import com.swirlds.merkledb.config.MerkleDbConfig_;
-import com.swirlds.virtualmap.config.VirtualMapConfig;
-import org.hiero.base.crypto.config.CryptoConfig;
-import org.hiero.consensus.config.BasicConfig;
-import org.hiero.consensus.config.PathsConfig;
-import org.hiero.consensus.reconnect.config.ReconnectConfig;
-import org.hiero.consensus.state.config.StateConfig;
-
-public final class ConfigUtils {
-    private ConfigUtils() {}
-
-    public static final Configuration CONFIGURATION = ConfigurationBuilder.create()
-            .withConfigDataType(BasicConfig.class)
-            .withConfigDataType(MerkleDbConfig.class)
-            .withSource(new SimpleConfigSource().withValue(MerkleDbConfig_.INITIAL_CAPACITY, "" + 65_536L))
-            .withConfigDataType(VirtualMapConfig.class)
-            .withConfigDataType(PathsConfig.class)
-            .withConfigDataType(CryptoConfig.class)
-            .withConfigDataType(StateConfig.class)
-            .withConfigDataType(PathsConfig.class)
-            .withConfigDataType(ReconnectConfig.class)
-            .build();
-}
