--- platform-sdk/swirlds-platform-core/src/testFixtures/java/module-info.java
@@ -1,30 +1,0 @@
-// SPDX-License-Identifier: Apache-2.0
-open module com.swirlds.platform.core.test.fixtures {
-    requires transitive com.hedera.node.hapi;
-    requires transitive com.swirlds.base;
-    requires transitive com.swirlds.common.test.fixtures;
-    requires transitive com.swirlds.common;
-    requires transitive com.swirlds.config.api;
-    requires transitive com.swirlds.config.extensions.test.fixtures;
-    requires transitive com.swirlds.platform.core;
-    requires transitive com.swirlds.state.api;
-    requires transitive com.swirlds.state.impl;
-    requires transitive com.swirlds.virtualmap;
-    requires transitive org.hiero.base.crypto;
-    requires transitive org.hiero.base.utility;
-    requires transitive org.hiero.consensus.model;
-    requires transitive org.hiero.consensus.platformstate;
-    requires com.swirlds.config.extensions;
-    requires com.swirlds.merkledb;
-    requires org.hiero.base.crypto.test.fixtures;
-    requires org.hiero.base.utility.test.fixtures;
-    requires org.hiero.consensus.reconnect;
-    requires org.hiero.consensus.state;
-    requires org.hiero.consensus.utility;
-    requires com.github.spotbugs.annotations;
-    requires org.junit.jupiter.api;
-    requires org.mockito;
-
-    exports com.swirlds.platform.test.fixtures.config;
-    exports com.swirlds.platform.test.fixtures.state;
-}
